% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/purityA-av-spectra.R
\docType{methods}
\name{averageFragmentation,purityA-method}
\alias{averageFragmentation,purityA-method}
\alias{frag4feature}
\title{Average fragemntation spectra across XCMS features}
\usage{
\S4method{averageFragmentation}{purityA}(pa, minfrac_intra = 0.5,
  minfrac_inter = 0.5, minnum_intra = 1, minnum_inter = 1,
  ppm_intra = 5, ppm_inter = 5, av = "median", sum_i = TRUE,
  purity_thres = 0.5)
}
\arguments{
\item{pa}{object; purityA object}

\item{xset}{object; XCMS object derived from the same files as the puritydf}

\item{ppm}{numeric; ppm tolerance between precursor mz and feature mz}

\item{plim}{numeric; min purity of precursor to be included}

\item{intense}{boolean; If the most intense precursor or the centered precursor is used}

\item{use_group}{boolean; Ignore individual peaks and just find matching fragmentation spectra within the (full) rtmin rtmax of each grouped feature}

\item{convert2RawRT}{boolean; If retention time correction has been used in XCMS set this to TRUE}

\item{create_db}{boolean; SQLite database will be created of the results}

\item{db_name}{character; If create_db is TRUE, a custom database name can be used, default is a time stamp}

\item{out_dir}{character; Path where database will be created}

\item{grp_peaklist}{dataframe [optional]; Can use any peak dataframe to add to databse. Still needs to be derived from the xset object though}
}
\value{
purityA object with slots for fragmentation-XCMS links
}
\description{
Average fragemntation spectra across XCMS features.
}
\examples{

msmsPths <- list.files(system.file("extdata", "lcms", "mzML", package="msPurityData"), full.names = TRUE, pattern = "MSMS")
xset <- xcms::xcmsSet(msmsPths, nSlaves = 1)
xset <- xcms::group(xset)
xset <- xcms::retcor(xset)
xset <- xcms::group(xset)

pa  <- purityA(msmsPths, interpol = "linear")
pa <- frag4feature(pa, xset)
pa <- averageFragmentation(pa)

}
