% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_database.R
\name{create_database}
\alias{create_database}
\title{Create database}
\usage{
create_database(pa, xset, out_dir, grp_peaklist = NA, db_name = NA)
}
\arguments{
\item{pa}{purityA object; Needs to be the same used for frag4feature function}

\item{xset}{xcms object; Needs to be the same used for frag4feature function}

\item{out_dir}{character; Out directory for the SQLite result database}

\item{grp_peaklist}{dataframe [optional]; Can use any peak dataframe. Still needs to be derived from the xset object though}

\item{db_name}{character [optional]; Name of the result database}
}
\value{
path to SQLite database
}
\description{
Create and SQLite database of an LC-MS(/MS) experiment
}
