% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/purityPD-dims-purity.R
\docType{methods}
\name{dimsPredictPurity,purityPD-method}
\alias{dimsPredictPurity}
\alias{dimsPredictPurity,purityPD-method}
\title{Using purityPD object, predict the precursor purity of processed peaks}
\usage{
\S4method{dimsPredictPurity}{purityPD}(Object, ppm = 1.5, minOffset = 0.5,
  maxOffset = 0.5, iwNorm = FALSE, iwNormFun = NULL, ilim = 0.05)
}
\arguments{
\item{Object}{object = purityPD object}

\item{ppm}{numeric = tolerance for target mz value in each scan}

\item{minOffset}{numeric = isolation window minimum offset}

\item{maxOffset}{numeric = isolation window maximum offset}

\item{iwNorm}{boolean = if TRUE then the intensity of the isolation window will be normalised based on the iwNormFun function}

\item{iwNormFun}{function = A function to normalise the isolation window intensity. The default function is very generalised and just accounts for edge effects}

\item{ilim}{numeric = All peaks less than this percentage of the target peak will be removed from the purity calculation, default is 5\% (0.05)}
}
\value{
purityPD object with predicted purity of peaks

purityPD object
}
\description{
Uses a purityPD object with references to multiple MS files. Predicts the
purity of the processed sample files
}
\examples{

datapth <- system.file("extdata", "dims", "mzML", package="msPurityData")
inDF <- Getfiles(datapth, pattern=".mzML", check = FALSE, cStrt = FALSE)
ppDIMS <- purityPD(fileList=inDF, cores=1, mzML=TRUE)
ppDIMS <- averageSpectra(ppDIMS)
ppDIMS <- filterp(ppDIMS)
ppDIMS <- subtract(ppDIMS)
ppDIMS <- dimsPredictPurity(ppDIMS)
}
\seealso{
\code{\link{dimsPredictPuritySingle}}
}

