% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_extract.R
\name{get_additional_mzml_meta}
\alias{get_additional_mzml_meta}
\title{Get additional mzML meta}
\usage{
get_additional_mzml_meta(mzml_pth)
}
\arguments{
\item{mzml_pth}{character; mzML path}
}
\value{
dataframe of meta info
}
\description{
Extract the filter strings 'accession MS:1000512' from an mzML file. Called header in thermo software.
Enables quick access to various information regarding each scan
}
\examples{
mzml_pth <- system.file("extdata", "dims", "mzML", 'B02_Daph_TEST_pos.mzML', package="msPurityData")
meta_df <- get_additional_mzml_meta(mzml_pth)
}
