% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/purityD-fileList.R
\name{Getfiles}
\alias{Getfiles}
\title{Get files for DI-MS processing}
\usage{
Getfiles(projectFolder = NULL, recursive = FALSE, pattern = ".csv",
  check = TRUE, raw = FALSE, peakout = NA, cStrt = TRUE,
  mzml_out = FALSE)
}
\arguments{
\item{projectFolder}{character: directory path}

\item{recursive}{boolean: recursively check for files}

\item{pattern}{character file suffix to check for}

\item{check}{boolean check with a GUI the files}

\item{raw}{(REDUNDANT)}

\item{peakout}{(REDUNDANT)}

\item{cStrt}{boolean use the first word as the class name for files}

\item{mzml_out}{(REDUNDANT)}
}
\value{
dataframe of files
}
\description{
Takes in a folder path and outputs the a data frame structure for purityD.
Function modified from mzmatch.
}
\examples{

datapth <- system.file("extdata", "dims", "mzML", package="msPurityData")
inDF <- Getfiles(datapth, pattern=".mzML", check = FALSE, cStrt = FALSE)
}

