% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/purityPD-class.R
\docType{methods}
\name{getP,purityPD-method}
\alias{getP}
\alias{getP,purityPD-method}
\title{Get peaklist for a purityPD object}
\usage{
\S4method{getP}{purityPD}(x)
}
\arguments{
\item{x}{object = purityPD object}
}
\value{
peaks
}
\description{
output peak list for a purityPD object
}
\examples{
datapth <- system.file("extdata", "dims", "mzML", package="msPurityData")
inDF <- Getfiles(datapth, pattern=".mzML", check = FALSE, cStrt = FALSE)
ppDIMS <- purityPD(fileList=inDF, cores=1, mzML=TRUE)
peaks <- getP(ppDIMS)
}

