% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/targeted-match-ions.R
\name{matchTargetIons}
\alias{matchTargetIons}
\title{Match list of m/z and/or RT values to msPurity grouped MS/MS spectral data}
\usage{
matchTargetIons(pa, ppm = 5, rtwin = 10, massTable, mzColumn = "mz",
  rtColumn = "rt")
}
\arguments{
\item{pa}{msPurity object}

\item{ppm}{Mass error range in ppm}

\item{rtwin}{RT window in seconds to use for feature matching}

\item{massTable}{Data frame containing list of masses to be matched.}

\item{mzColumn}{Name of the column in massTable for m/z values}

\item{rtColumn}{Name of the column in massTable for RT values. If set to NULL full RT range os measurement will be used}
}
\description{
Match list of m/z and/or RT values to msPurity grouped MS/MS spectral data
}
