% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msPurity.R
\docType{package}
\name{msPurity-package}
\alias{msPurity}
\alias{msPurity-package}
\title{\code{msPurity} package}
\description{
\href{https://bioconductor.org/packages/release/bioc/html/msPurity.html}{msPurity Bioconductor}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/computational-metabolomics/msPurity/}
  \item Report bugs at \url{https://github.com/computational-metabolomics/msPurity/issues/new}
}

}
\author{
\strong{Maintainer}: Thomas N. Lawson \email{thomas.nigel.lawson@gmail.com} (\href{https://orcid.org/0000-0002-5915-7980}{ORCID})

Authors:
\itemize{
  \item Thomas N. Lawson \email{thomas.nigel.lawson@gmail.com} (\href{https://orcid.org/0000-0002-5915-7980}{ORCID})
}

Other contributors:
\itemize{
  \item Ralf Weber [contributor]
  \item Martin Jones [contributor]
  \item Julien Saint-Vanne [contributor]
  \item Andris Jankevics [contributor]
  \item Ossama Edbali \email{o.edbali@bham.ac.uk} (\href{https://orcid.org/0000-0003-0132-8668}{ORCID}) [contributor]
  \item Mark Viant [thesis advisor]
  \item Warwick Dunn [thesis advisor]
}

}
\keyword{internal}
