% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/targeted-match-ions.R
\name{mzRange}
\alias{mzRange}
\title{Calculate m/z range for given m/z value and ppm range}
\usage{
mzRange(ppm, mass)
}
\arguments{
\item{ppm}{Mass error range in ppm}

\item{mass}{m/z value for which to calculate range}
}
\description{
Calculate m/z range for given m/z value and ppm range
}
