% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/purityPD-class.R
\docType{class}
\name{purityPD-class}
\alias{purityPD}
\alias{purityPD-class}
\title{An S4 class to represent a DI-MS purityPD}
\arguments{
\item{.Object}{object = purityPD object}

\item{fileList}{data.frame = created using GetFiles, data.frame with filepaths and sample class information}

\item{cores}{numeric = Number of cores used to perform Hierarchical clustering WARNING: memory intensive, default 1}

\item{mzML}{boolean = TRUE if mzML to be used FALSE if .csv file to be used}
}
\value{
purityPD object
}
\description{
The class used to predict purity from an DI-MS dataset.
}
\examples{
datapth <- system.file("extdata", "dims", "mzML", package="msPurityData")
inDF <- Getfiles(datapth, pattern=".mzML", check = FALSE, cStrt = FALSE)
ppDIMS <- purityPD(fileList=inDF, cores=1, mzML=TRUE)
}

