% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/purityPL-constructor.R
\name{purityPL}
\alias{purityPL}
\title{Predict the precursor purity from an LC-MS run}
\usage{
purityPL(xset, purityType = "purityFWHMmedian", offsets = c(0.5, 0.5),
  fileignore = NULL, cores = 1, xgroups = NULL, iwNorm = FALSE,
  iwNormFun = NULL, ilim = 0)
}
\arguments{
\item{xset}{object = xcms object}

\item{purityType}{character = Area and average used for the purity predictions. Options are
"purityFWHMmedian", "purityFWmedian", "purityFWHMmean", "purityFWmean"}

\item{offsets}{vector = vector of the isolation window upper and lower offsets}

\item{fileignore}{vector = vector of files to ignore for the prediction calculation}

\item{cores}{numeric = number of cores to use}

\item{xgroups}{vector = vector of xcms groups to perform prediction on}

\item{iwNorm}{boolean = if TRUE then the intensity of the isolation window will be normalised based on the iwNormFun function}

\item{iwNormFun}{function = A function to normalise the isolation window intensity. The default function is very generalised and just accounts for edge effects}

\item{ilim}{numeric = All peaks less than this percentage of the target peak will be removed from the purity calculation, default is 5\% (0.05)}
}
\value{
a purityPL object containing a dataframe of predicted purity scores
}
\description{
Constructor for the purityLC class.

Given an XCMS object predict the purity of the grouped peaks
}
\examples{
msPths <- list.files(system.file("extdata", "lcms", "mzML", package="msPurityData"), full.names = TRUE, pattern = "LCMS_")
xset <- xcms::xcmsSet(msPths)
xset <- xcms::group(xset)
xset <- xcms::retcor(xset)
xset <- xcms::group(xset)
ppLCMS <- purityPL(xset, cores = 1, xgroups = c(1, 2))

}

