% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/purityA-0-class.R
\name{show,purityA-method}
\alias{show,purityA-method}
\title{Show method for purityA class}
\usage{
\S4method{show}{purityA}(object)
}
\arguments{
\item{object}{object; purityA object}
}
\value{
a print statement of regarding object
}
\description{
print statement for purityA class
}
