% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/purityD-class.R
\docType{methods}
\name{show,purityD-method}
\alias{show,purityD-method}
\title{Show method for purityD}
\usage{
\S4method{show}{purityD}(object)
}
\arguments{
\item{object}{= purityD object}
}
\value{
a print statement of regarding object
}
\description{
Show method for purityD object
}

