% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/purityPD-class.R
\docType{methods}
\name{show,purityPD-method}
\alias{show,purityPD-method}
\title{Show method for purityPD}
\usage{
\S4method{show}{purityPD}(object)
}
\arguments{
\item{object}{= purityPD object}
}
\value{
a print statement of regarding object
}
\description{
Show method for purityPD object
}

