% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/purityPL-class.R
\docType{methods}
\name{show,purityPL-method}
\alias{show,purityPL-method}
\title{Show method for purityPL}
\usage{
\S4method{show}{purityPL}(object)
}
\arguments{
\item{object}{= purityPL object}
}
\value{
a print statement of regarding object
}
\description{
Show method for purityPL object
}

