% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectral_matching.R
\name{spectral_matching}
\alias{spectral_matching}
\title{Spectral matching}
\usage{
spectral_matching(query_db_pth, ra_thres_l = 0, ra_thres_q = 2,
  cores = 1, pol = "positive", ppm_tol_prod = 10, ppm_tol_prec = 5,
  score_thres = 0.6, topn = NA, db_name = NA, library_db_pth = NA,
  instrument_types = NA, library_sources = "massbank", scan_ids = NA,
  pa = NA, xset = NA, grp_peaklist = NA, out_dir = ".",
  ra_w_q = 0.5, ra_w_l = 0.5, mz_w_q = 2, mz_w_l = 2,
  spectra_type_q = "scans", ra_thres_t = NA, target_db_pth = NA)
}
\arguments{
\item{query_db_pth}{character; Path of the database of targets (queries) that will be searched against the library spectra. Generated
either from frag4feature or from create_database functions.}

\item{ra_thres_l}{numeric; Relative abundance threshold for library spectra}

\item{ra_thres_q}{numeric; Relative abundance threshold for target (query) spectra
(Peaks below this RA threshold will be excluded)}

\item{cores}{numeric; Number of cores to use}

\item{pol}{character; Polarity ['positive' or 'negative']}

\item{ppm_tol_prod}{numeric; PPM tolerance to match to product}

\item{ppm_tol_prec}{numeric; PPM tolerance to match to precursor}

\item{score_thres}{numeric; Dot product cosine score threshold}

\item{topn}{numeric [optional]; Only use top n matches}

\item{db_name}{character [optional]; Name of the result database
(e.g. can use CAMERA peaklist)}

\item{library_db_pth}{character [optional]; path to library spectral SQLite database.
Defaults to msPurityData package data.}

\item{instrument_types}{vector [optional]; Vector of instrument types, defaults to all}

\item{library_sources}{vector [optional]; Vector of library sources. Default option is for massbank only but the 'lipidblast'
library is also available}

\item{scan_ids}{vector [optional]; Vector of unique scan ids calculated from msPurity "pid". These scans will on
used for the spectral matching. All scans will be used if set to NA}

\item{pa}{purityA object [optional]; If target_db_pth set to NA, a new database can be created using pa, xset and grp_peaklist}

\item{xset}{xcms object [optional]; If target_db_pth set to NA, a new database can be created using pa, xset and grp_peaklist}

\item{grp_peaklist}{dataframe [optional]; If target_db_pth set to NA, a new database can be created using pa, xset and grp_peaklist}

\item{out_dir}{character [optional]; If target_db_pth set to NA, Out directory for the SQLite result database}

\item{ra_w_q}{numeric; Relative abundance weight for query spectra}

\item{ra_w_l}{numeric; Relative abundance weight for library spectra}

\item{mz_w_q}{numeric; mz weight for query spectra}

\item{mz_w_l}{numeric; mz weight for library spectra}

\item{spectra_type_q}{character; Type of fragmentation spectra from query to match with "scans" =  all individual scans,
"av_intra" = averaged spectra (intra), "av_inter" = averaged spectra (inter), "av_all" = averaged all
 spectra ignoring inter-intra relationships}

\item{ra_thres_t}{numeric [deprecated]; The relative abundance threshold for the query spectra (use ra_thres_q for future use)}

\item{target_db_pth}{character [deprecated]; The query database path (use query_db_pth for future use)}
}
\value{
list of database details and dataframe summarising the results for the xcms features
}
\description{
Perform spectral matching to spectral libraries using dot product cosine on a LC-MS/MS dataset and link to XCMS features.
}
\examples{
msmsPths <- list.files(system.file("extdata", "lcms", "mzML", package="msPurityData"), full.names = TRUE, pattern = "MSMS")
xset <- xcms::xcmsSet(msmsPths, nSlaves = 1)
xset <- xcms::group(xset)
xset <- xcms::retcor(xset)
xset <- xcms::group(xset)

pa  <- purityA(msmsPths)
pa <- frag4feature(pa, xset, create_db=TRUE)
#NOTE that scan_ids here are refer the unique scan id calculated by purityA (pids).
#Only required if you want to limit the spectral matching to certain scans
result <- spectral_matching(pa@db_path, scan_ids = c(1120,  366, 1190, 601,  404,1281, 1323, 1289))
}
