% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/purityA-validate.R
\name{validate,purityA-method}
\alias{validate,purityA-method}
\alias{validate}
\title{Validate precursor purity predictions using LC-MS and LC-MS/MS dataset}
\usage{
\S4method{validate}{purityA}(pa, ppLCMS)
}
\arguments{
\item{pa}{object; purityA object}

\item{ppLCMS}{object; purityX object}
}
\value{
purityA object
}
\description{
The method is used to validate the precursor purity predictions made
from an LC-MS dataset
}
