% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/purityD-writeOut.R
\docType{methods}
\name{writeOut,purityD-method}
\alias{writeOut,purityD-method}
\alias{writeOut}
\title{Using purityD object, save peaks as text files}
\usage{
\S4method{writeOut}{purityD}(Object, outDir, original)
}
\arguments{
\item{Object}{object = purityD object}

\item{outDir}{character = Directory to save text files}

\item{original}{boolean = if the original (unprocessed) files are to be saved to text files}
}
\value{
purityD object
}
\description{
Uses a purityD object with references to multiple MS files.
Predicts the purity of the processed sample files
}
