% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/purityPD-writeOut.R
\docType{methods}
\name{writeOut,purityPD-method}
\alias{writeOut}
\alias{writeOut,purityPD-method}
\title{Using purityPD object, save peaks as text files}
\usage{
\S4method{writeOut}{purityPD}(Object, outDir, original)
}
\arguments{
\item{Object}{object = purityPD object}

\item{outDir}{character = Directory to save text files}

\item{original}{boolean = if the original (unprocessed) files are to be saved to text files}
}
\value{
purityPD object
}
\description{
Uses a purityPD object with references to multiple MS files.
Predicts the purity of the processed sample files
}

