% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/purityA-av-spectra.R
\docType{methods}
\name{averageAllFragSpectra,purityA-method}
\alias{averageAllFragSpectra,purityA-method}
\alias{averageAllFragSpectra}
\title{Average and filter fragmentation spectra for each XCMS feature within and accross MS data files (ignoring intra and inter relationships)}
\usage{
\S4method{averageAllFragSpectra}{purityA}(pa, minfrac = 0.5,
  minnum = 1, ppm = 5, snr = 0, ra = 0, snr_pre = 0,
  ra_pre = 0, av = "median", sum_i = TRUE, plim = 0.5,
  remove_peaks = FALSE, cores = 1)
}
\arguments{
\item{pa}{object; purityA object}

\item{minfrac}{numeric;minimum ratio of the peak fraction (peak count / total peaks) across all (ignoring intra and inter relationships)}

\item{minnum}{numeric; minimum number of times peak is present across all fragmentation spectra (ignoring intra and inter relationships)}

\item{ppm}{numeric; ppm threshold to average across all scans (ignoring intra and inter relationships)}

\item{snr}{numeric;  minimum signal-to-noise of the peak across all (ignoring intra and inter relationships)}

\item{ra}{numeric; minimum relative abundance of the peak fraction across all (ignoring intra and inter relationships)}

\item{snr_pre}{numeric;  minimum signal-to-noise prior to averaging}

\item{ra_pre}{numeric;  minimum relative abundance prior to averaging}

\item{av}{character; type of averaging to use (median or mean)}

\item{sum_i}{boolean; TRUE if the intensity for each peak is summed across averaged spectra}

\item{plim}{numeric; min purity of precursor for fragmentation spectra scan to be included}

\item{remove_peaks}{boolean; TRUE if peaks are to be removed that do not meet the threshold criteria. Otherwise they will just be flagged}

\item{cores}{numeric; Number of cores for multiprocessing}
}
\description{
Average and filter fragmentation spectra for each XCMS feature within and accross MS data files (ignoring intra and inter relationships).

The default approach is to use hierarchical clustering where peaks within a set ppm tolerance will be clustered.

The clustered peaks are then averaged (or summed) and filtered.
}
\examples{

msmsPths <- list.files(system.file("extdata", "lcms", "mzML", package="msPurityData"), full.names = TRUE, pattern = "MSMS")
xset <- xcms::xcmsSet(msmsPths, nSlaves = 1)
xset <- xcms::group(xset)
xset <- xcms::retcor(xset)
xset <- xcms::group(xset)

pa  <- purityA(msmsPths, interpol = "linear")
pa <- frag4feature(pa, xset)
pa <- averageAllFragSpectra(pa)

}
