% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/purityA-av-spectra.R
\docType{methods}
\name{averageFragmentation,purityA-method}
\alias{averageFragmentation,purityA-method}
\alias{averageFragmentation}
\title{Average fragemntation spectra across XCMS features}
\usage{
\S4method{averageFragmentation}{purityA}(pa, minfrac_intra = 0.5,
  minfrac_inter = 0.5, minnum_intra = 1, minnum_inter = 1,
  ppm_intra = 5, ppm_inter = 5, minfrac_all = 0.5, minnum_all = 1,
  ppm_all = 5, av = "median", sum_i = TRUE, plim = 0.5,
  remove_peaks = FALSE, cores = 1)
}
\arguments{
\item{pa}{object; purityA object}

\item{minfrac_intra}{numeric; minimum ratio of the peak fraction (peak count / total peaks) within each file}

\item{minfrac_inter}{numeric; minimum ratio of the peak fraction (peak count / total peaks) across files}

\item{minnum_intra}{numeric; minimum number of times peak is present across fragmentation spectra within each file}

\item{minnum_inter}{numeric; minimum number of times peak is present across fragmentation spectra across files}

\item{ppm_intra}{numeric; ppm threshold to average within each file}

\item{ppm_inter}{numeric; ppm threshold to average across files}

\item{minfrac_all}{numeric;minimum ratio of the peak fraction (peak count / total peaks) across all (ignoring intra and inter relationships)}

\item{minnum_all}{numeric; minimum number of times peak is present across all fragmentation spectra (ignoring intra and inter relationships)}

\item{ppm_all}{numeric; ppm threshold to average across all scans (ignoring intra and inter relationships)}

\item{av}{character; type of averaging to use (median or mean)}

\item{sum_i}{boolean; TRUE if the intensity for each peak is summed across averaged spectra}

\item{plim}{numeric; min purity of precursor for fragmentation spectra scan to be included}

\item{remove_peaks}{boolean; TRUE if peaks are to be removed that do not meet the minfrac criteria. Otherwise they will just be flagged}

\item{cores}{numeric; Number of cores for multiprocessing}
}
\description{
Average fragemntation spectra across XCMS features. Averaging can be done within file (intra), across files (inter)
and independantly of the files (all).

The default approach is to use hierarchical clustering where peaks within a set ppm tolerance will be clustered.

The clustered peaks are then averaged (or summed) and filtered.
}
\examples{

msmsPths <- list.files(system.file("extdata", "lcms", "mzML", package="msPurityData"), full.names = TRUE, pattern = "MSMS")
xset <- xcms::xcmsSet(msmsPths, nSlaves = 1)
xset <- xcms::group(xset)
xset <- xcms::retcor(xset)
xset <- xcms::group(xset)

pa  <- purityA(msmsPths, interpol = "linear")
pa <- frag4feature(pa, xset)
pa <- averageFragmentation(pa)

}
