% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/purityA-av-spectra.R
\docType{methods}
\name{averageIntraFragSpectra,purityA-method}
\alias{averageIntraFragSpectra,purityA-method}
\alias{averageIntraFragSpectra}
\title{Average and filter fragmentation spectra for each XCMS feature within a MS data file}
\usage{
\S4method{averageIntraFragSpectra}{purityA}(pa, minfrac = 0.5,
  minnum = 1, ppm = 5, snr = 0, ra = 0, av = "median",
  sumi = TRUE, rmp = FALSE, cores = 1)
}
\arguments{
\item{pa}{object; purityA object}

\item{minfrac}{numeric; minimum ratio of the peak fraction (peak count / total peaks) within each file}

\item{minnum}{numeric; minimum number of times peak is present across fragmentation spectra within each file}

\item{ppm}{numeric; ppm threshold to average within each file}

\item{snr}{numeric; minimum signal-to-noise of the peak within each file}

\item{ra}{numeric; minimum relative abundance of the peak within each file}

\item{av}{character; type of averaging to use (median or mean)}

\item{sumi}{boolean; TRUE if the intensity for each peak is summed across averaged spectra}

\item{rmp}{boolean; TRUE if peaks are to be removed that do not meet the threshold criteria. Otherwise they will just be flagged}

\item{cores}{numeric; Number of cores for multiprocessing}
}
\description{
Average and filter fragmentation spectra for each XCMS feature within a MS data file.

The default approach is to use hierarchical clustering where peaks within a set ppm tolerance will be clustered.

The clustered peaks are then averaged (or summed) and filtered.
}
\examples{

msmsPths <- list.files(system.file("extdata", "lcms", "mzML", package="msPurityData"), full.names = TRUE, pattern = "MSMS")
xset <- xcms::xcmsSet(msmsPths, nSlaves = 1)
xset <- xcms::group(xset)
xset <- xcms::retcor(xset)
xset <- xcms::group(xset)

pa  <- purityA(msmsPths, interpol = "linear")
pa <- frag4feature(pa, xset)
pa <- averageIntraFragSpectra(pa)

}
