% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/purityPD-av-spectra.R
\name{averageSpectraSingle}
\alias{averageSpectraSingle}
\title{Calculates to average mz, intensity and signal-to-noise of multiple
 scans from 1 MS datafile (mzML or .csv)}
\usage{
averageSpectraSingle(filePth, rtscn = "all", scanRange = NA,
  timeRange = NA, clustType = "hc", ppm = 1.5, snthr = 3, cores = 1,
  av = "median", missingV = "ignore", minfrac = 0.6667,
  snMeth = "median", MSFileReader = FALSE, normTIC = FALSE)
}
\arguments{
\item{filePth}{character = Path of the file to be processed}

\item{rtscn}{character = Whether it is scans or retention time to be filtered. Use "all" if all scans to be used. ['rt', 'scns', 'all']}

\item{scanRange}{vector = Scan range (if rtscn='scns') e.g. c(40, 69)}

\item{timeRange}{vector = Time range (if rtscn='rt) e.g. c(10.3, 400.8) (only if using mzML file)}

\item{clustType}{character = Type of clustering used either Hierarchical or just simple 1dgrouping ['hc', 'simple'], default 'hc'}

\item{ppm}{numeric = the ppm error to cluster mz together default 1.5}

\item{snthr}{numeric = Signal to noise ratio threshold, default 0}

\item{cores}{numeric = Number of cores used to perform Hierarchical clustering WARNING: memory intensive, default 2}

\item{av}{character = What type of averaging to do between peaks}

\item{missingV}{character = What to do with missing values (zero or ignore)}

\item{minfrac}{numeric = Min fraction of scans with a grouped peak to be an accepted averaged peak}

\item{snMeth}{character = Type of snMethod to use}

\item{MSFileReader}{boolean = For thermo files a the MSFileReader API can extract peaklist. This can consist of an .csv file with
the following columns c('mz', 'i', 'scanid', 'snr')}

\item{normTIC}{boolean = If TRUE then RSD calculation will use the normalised intensity (intensity divided by TIC) if FALSE will use standard intensity}
}
\value{
dataframe of the median mz, intensity, signal-to-noise ratio.
}
\description{
Averages multiple scans of mass spectrometry data together.
Each scan consisting of a minimum of intensity and mz values.

Works for either mzML or a .csv file consisting of mz, i, scanid,
 (optional: noise, backgroun, snr)

Signal-to-noise (SNR) can be calculated a number of ways. Default is to
calculate the SN for every scan as the
"Intensity of peak / the median intensity of the scan".

Alternatively if using a .CSV file a precalculated snr can be on of the
columns and this can be used.

The function works for LC-MS or DI-MS datasets.
}
\examples{
mzmlPth <- system.file("extdata", "dims", "mzML", "B02_Daph_TEST_pos.mzML", package="msPurityData")
avP <- averageSpectraSingle(mzmlPth)
}

