% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_annotations.R
\name{combineAnnotations}
\alias{combineAnnotations}
\title{Combine Annotations}
\usage{
combineAnnotations(sqlitePth, metfrag_resultPth = NA,
  sirius_csi_resultPth = NA, probmetab_resultPth = NA,
  lipidsearch_resultPth = NA, mzcloud_resultPth = NA, xset,
  weights = c(0.2, 0.1, 0.1, 0.05, 0.2, 0.3, 0.05), avFrag = "av_all")
}
\description{

}
\examples{

}
