% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combineAnnotations.R
\name{combineAnnotations}
\alias{combineAnnotations}
\title{Combine Annotations}
\usage{
combineAnnotations(sm_resultPth, metfrag_resultPth = NA,
  sirius_csi_resultPth = NA, probmetab_resultPth = NA,
  weights = list(sm = 0.4, metfrag = 0.25, sirius_csifingerid = 0.25,
  probmetab = 0.1), outPth = NA, silentRestErrors = FALSE)
}
\arguments{
\item{sm_resultPth}{character;}

\item{metfrag_resultPth}{character;}

\item{sirius_csi_resultPth}{character;}

\item{probmetab_resultPth}{character;}

\item{weights}{list;}

\item{outPth}{character;}

\item{silentRestErrors}{boolean;}
}
\value{
purityA object with slots for fragmentation-XCMS links
}
\description{
Combine the annotation results from msPurity spectral matching, MetFrag, Sirius CSI:FingerID and probmetab
 based on weighted scores for each technique aligning each annotation by inchikey and XCMS grouped feature.
}
\examples{
metfrag_resultPth <- system.file("extdata", "external_annotations", "metfrag.tsv", package="msPurity")
# run the standard spectral matching workflow to get the sm_resultPth
sm_resultPth <- system.file("extdata","tests", "sm", "spectralMatching_result.sqlite", package="msPurity")

combined <- combineAnnotations(sm_resultPth, metfrag_resultPth, outPth=file.path(tempdir(), 'combined.sqlite'))
}
