% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_annotations.R
\name{combineAnnotations}
\alias{combineAnnotations}
\title{Combine Annotations}
\usage{
combineAnnotations(sqlitePth, metfrag_resultPth = NA,
  sirius_csi_resultPth = NA, probmetab_resultPth = NA,
  weights = list(sm = 0.4, metfrag = 0.2, sirius_csifingerid = 0.2,
  probmetab = 0.1), silentRestErrors = FALSE)
}
\arguments{
\item{sqlitePth}{character;}

\item{metfrag_resultPth}{character;}

\item{sirius_csi_resultPth}{character;}

\item{probmetab_resultPth}{character;}

\item{weights}{list;}

\item{silentRestErrors}{boolean;}
}
\value{
purityA object with slots for fragmentation-XCMS links
}
\description{
Combine the annotation results from msPurity spectral matching, MetFrag, Sirius CSI:FingerID and probmetab
 based on weighted scores for each technique aligning each annotation by inchikey and XCMS grouped feature.
}
\examples{
metfrag_resultPth <- system.file("extdata", "external_annotations", "metfrag.tsv", package="msPurity")
sirius_csi_resultPth <- system.file("extdata", "external_annotations", "sirus_csifingerid.tsv", package="msPurity")
probmetab_resultPth <- system.file("extdata", "external_annotations", "probmetab.tsv", package="msPurity")

msmsPths <- list.files(system.file("extdata", "lcms", "mzML", package="msPurityData"), full.names = TRUE, pattern = "MSMS")
xset <- xcms::xcmsSet(msmsPths)
xset <- xcms::group(xset)

pa  <- purityA(msmsPths)
pa <- frag4feature(pa, xset)
pa <- averageAllFragSpectra(pa)
dbpth <- create_database(pa, xset)
result <- spectral_matching(dbpth, spectra_type_q="av_all")
combineAnnotations(dbpth, metfrag_resultPth, sirius_csi_resultPth, probmetab_resultPth)
}
