% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/purityA-create-msp.R
\docType{methods}
\name{createMSP,purityA-method}
\alias{createMSP,purityA-method}
\alias{createMSP}
\title{Create MSP file from purityA object}
\usage{
\S4method{createMSP}{purityA}(pa, msp_file_pth = NULL, metadata = NULL,
  metadata_cols = c("CH$NAME", "MS$FOCUSED_ION: PRECURSOR_TYPE"),
  xcms_groupids = NULL, method = "all", adduct_split = TRUE)
}
\arguments{
\item{pa}{object;  purityA object}

\item{msp_file_pth}{character; Name of the output msp file, if NULL the file "frag_spectra_{time stamp}.msp" will be created in the current directory}

\item{metadata}{data.frame; Data frame with additional coumpound infomation to include in msp output}

\item{metadata_cols}{vector; Column names of meta data to incorporate into name}

\item{xcms_groupids}{vector; XCMS group id's to extract ms/ms data for}

\item{method}{character; "all" will export all matching ms/ms spectra to xcms features, "max" will use spectra with the highest inensity,
"av-intra" will use the intra file averaged file, "av-inter" will use the inter file (within file)
 averaged spectra, "av-all" will use the averaged spectra (ignoring inter and intra)}

\item{adduct_split}{boolean; If either "adduct" or  MS$FOCUSED_ION: PRECURSOR_TYPE column is in metadata then each adduct will have it's own MSP spectra.
(Useful, if the MSP file will be used for further annotation)}
}
\description{
Create an MSP file for all the fragmentation spectra that has been linked to an XCMS feature via frag4feature
}
\examples{

msmsPths <- list.files(system.file("extdata", "lcms", "mzML", package="msPurityData"), full.names = TRUE, pattern = "MSMS")
xset <- xcms::xcmsSet(msmsPths, nSlaves = 1)
xset <- xcms::group(xset)
xset <- xcms::retcor(xset)
xset <- xcms::group(xset)

pa  <- purityA(msmsPths)
pa <- frag4feature(pa, xset)
pa <- averageFragmentation(pa)
createMSP(pa)
}
