% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/purityA-filter-frag-spectra.R
\docType{methods}
\name{filterFragSpectra,purityA-method}
\alias{filterFragSpectra,purityA-method}
\title{Filter fragmentations spectra associated with an XCMS feature}
\usage{
\S4method{filterFragSpectra}{purityA}(pa, i = 0, p = 0.8, ra = 0,
  snr = 3, rmp = FALSE, snmeth = "median")
}
\arguments{
\item{pa}{object; purityA object}

\item{ra}{numeric; minimum relative abundance of a peak}

\item{snr}{numeric; minimum signal-to-noise of a peak  peak within each file}

\item{cores}{numeric; Number of cores for multiprocessing}

\item{ilim}{numeric; min intensity of a peak}

\item{plim}{numeric; min precursor ion purity of the associated precursor for fragmentation spectra scan}

\item{remove_peaks}{boolean; TRUE if peaks are to be removed that do not meet the threshold criteria. Otherwise they will just be flagged}
}
\description{
Flag and filter features based on signal-to-noise ratio, relative abundance, intensity threshold and
precursor ion purity of precursor.

The grp_spectra slot to add the columns snr, ra, snr_flag, ra_flag, purity_flag, intensity_flag pass_flag

This filtering occurs at the scan level (i.e. should be run prior to any averaging)
}
\examples{

msmsPths <- list.files(system.file("extdata", "lcms", "mzML", package="msPurityData"), full.names = TRUE, pattern = "MSMS")
xset <- xcms::xcmsSet(msmsPths)
xset <- xcms::group(xset)
xset <- xcms::retcor(xset)
xset <- xcms::group(xset)

pa  <- purityA(msmsPths)
pa <- frag4feature(pa, xset)
pa <- filterFragSpectra(pa)


}
