% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/purityA-frag4feature.R
\name{frag4feature,purityA-method}
\alias{frag4feature,purityA-method}
\alias{frag4feature}
\title{Using a purityA object, link MS/MS data to XCMS features}
\usage{
\S4method{frag4feature}{purityA}(
  pa,
  xset,
  ppm = 5,
  plim = NA,
  intense = TRUE,
  convert2RawRT = TRUE,
  useGroup = FALSE,
  create_db = FALSE,
  out_dir = ".",
  db_name = NA,
  grp_peaklist = NA,
  use_group = NA
)
}
\arguments{
\item{pa}{object; purityA object}

\item{xset}{object; xcmsSet object derived from the same files as those used to create the purityA object}

\item{ppm}{numeric; ppm tolerance between precursor mz and XCMS feature mz}

\item{plim}{numeric; minimum purity of precursor to be included}

\item{intense}{boolean; If TRUE the most intense precursor will be used. If FALSE the precursor closest to the center of the isolation window will be used}

\item{convert2RawRT}{boolean; If retention time correction has been used in XCMS set this to TRUE}

\item{useGroup}{boolean; Ignore individual peaks and just find matching fragmentation spectra within the (full) rtmin rtmax of each grouped feature}

\item{create_db}{boolean; (Deprecated, to be removed - use createDatabase function) SQLite database will be created of the results}

\item{out_dir}{character; (Deprecated, to be removed - use createDatabase function) Path where database will be created}

\item{db_name}{character; (Deprecated, to be removed - use createDatabase function) If create_db is TRUE, a custom database name can be used, default is a time stamp}

\item{grp_peaklist}{dataframe; (Deprecated, to be removed - use createDatabase function) Can use any peak dataframe to add to databse. Still needs to be derived from the xset object though}

\item{use_group}{boolean; (Deprecated, to be removed - replaced with useGroup argument for style consistency)}
}
\value{
Returns a purityA object (pa) with the following slots populated:
\itemize{
\item pa@grped_df: A dataframe of the grouped XCMS features linked to the associated fragmentation spectra precursor details is recorded here
\item pa@grped_ms2: A list of fragmentation spectra associated with each grouped XCMS feature is recorded here
\item pa@f4f_link_type: The linking method is recorded here (e.g. individual peaks or grouped - "useGroup=TRUE")
}
}
\description{

}
\examples{

msmsPths <- list.files(system.file("extdata", "lcms",
                       "mzML", package="msPurityData"), full.names = TRUE,
                       pattern = "MSMS")
xset <- xcms::xcmsSet(msmsPths, nSlaves = 1)
xset <- xcms::group(xset)
xset <- xcms::retcor(xset)
xset <- xcms::group(xset)

pa  <- purityA(msmsPths)
pa <- frag4feature(pa, xset)
}
