% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/purityA-frag4feature.R
\docType{methods}
\name{frag4feature,purityA-method}
\alias{frag4feature,purityA-method}
\alias{frag4feature}
\title{Assign precursor purity scored fragmentation spectra to XCMS features}
\usage{
\S4method{frag4feature}{purityA}(pa, xset, ppm = 5, plim = NA,
  intense = TRUE, convert2RawRT = TRUE, create_db = FALSE,
  out_dir = ".", db_name = NA, grp_peaklist = NA,
  use_group = FALSE)
}
\arguments{
\item{pa}{object; purityA object}

\item{xset}{object; XCMS object derived from the same files as the puritydf}

\item{ppm}{numeric; ppm tolerance between precursor mz and feature mz}

\item{plim}{numeric; min purity of precursor to be included}

\item{intense}{boolean; If the most intense precursor or the centered precursor is used}

\item{convert2RawRT}{boolean; If retention time correction has been used in XCMS set this to TRUE}

\item{create_db}{boolean; SQLite database will be created of the results}

\item{out_dir}{character; Path where database will be created}

\item{db_name}{character; If create_db is TRUE, a custom database name can be used, default is a time stamp}

\item{grp_peaklist}{dataframe [optional]; Can use any peak dataframe to add to databse. Still needs to be derived from the xset object though}

\item{use_group}{boolean; Ignore individual peaks and just find matching fragmentation spectra within the (full) rtmin rtmax of each grouped feature}
}
\value{
purityA object with slots for fragmentation-XCMS links
}
\description{
Assign fragmentation spectra (MS/MS) scored via msPurity package to features
from an XCMS set object.

Allows the user to filter out spectra below a certain threshold for purity.
}
\examples{

msmsPths <- list.files(system.file("extdata", "lcms", "mzML", package="msPurityData"), full.names = TRUE, pattern = "MSMS")
xset <- xcms::xcmsSet(msmsPths, nSlaves = 1)
xset <- xcms::group(xset)
xset <- xcms::retcor(xset)
xset <- xcms::group(xset)

pa  <- purityA(msmsPths, interpol = "linear")
pa <- frag4feature(pa, xset)

}
