% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectral_matching.R
\name{spectral_matching}
\alias{spectral_matching}
\title{Spectral matching}
\usage{
spectral_matching(pa, xset, ra_thres_l = 0, ra_thres_t = 2, cores = 1,
  pol = "positive", ppm_tol_prod = 10, ppm_tol_prec = 5,
  score_thres = 0.6, out_dir = ".", topn = NA, db_name = NA,
  grp_peaklist = NA, library_db_pth = NA, instrument_types = NA,
  library_sources = "massbank", scan_ids = NA)
}
\arguments{
\item{pa}{purityA object; Needs to be the same used for frag4feature function}

\item{xset}{xcms object; Needs to be the same used for frag4feature function}

\item{ra_thres_l}{numeric; Relative abundance threshold for library spectra}

\item{ra_thres_t}{numeric; Relative abundance threshold for target spectra
(Peaks below this RA threshold will be excluded)}

\item{cores}{numeric; Number of cores to use}

\item{pol}{character; Polarity ['positive' or 'negative']}

\item{ppm_tol_prod}{numeric; PPM tolerance to match to product}

\item{ppm_tol_prec}{numeric; PPM tolerance to match to precursor}

\item{score_thres}{numeric; Dot product cosine score threshold}

\item{out_dir}{character; Out directory for the SQLite result database}

\item{topn}{numeric [optional]; Only use top n matches}

\item{db_name}{character [optional]; Name of the result database}

\item{grp_peaklist}{dataframe [optional]; Can use any peak dataframe. Still needs to be derived from the xset object though
(e.g. can use CAMERA peaklist)}

\item{library_db_pth}{character [optional]; path to library spectral SQLite database. Defaults to msPurityData package data.}

\item{instrument_types}{vector [optional]; Vector of instrument types, defaults to all}

\item{library_sources}{vector [optional]; Vector of library sources. Default option is for massbank only but the 'lipidblast'
library is also available}

\item{scan_ids}{vector [optional]; Vector of unique scan ids calculated from msPurity "pid". These scans will on
used for the spectral matching. All scans will be used if set to NA}
}
\value{
list of database details and dataframe summarising the results for the xcms features
}
\description{
Perform spectral matching to spectral libraries using dot product cosine on a LC-MS/MS dataset and link to XCMS features.
}
\examples{
msmsPths <- list.files(system.file("extdata", "lcms", "mzML", package="msPurityData"), full.names = TRUE, pattern = "MSMS")
xset <- xcms::xcmsSet(msmsPths, nSlaves = 1)
xset <- xcms::group(xset)
xset <- xcms::retcor(xset)
xset <- xcms::group(xset)

pa  <- purityA(msmsPths)
pa <- frag4feature(pa, xset)
#NOTE that scan_ids here are refer the unique scan id calculated by purityA (pids).
#Only required if you want to limit the spectral matching to certain scans
result <- spectral_matching(pa, xset, scan_ids = c(1120,  366, 1190, 601,  404,1281, 1323, 1289))
}
