% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectralMatching.R
\name{spectralMatching}
\alias{spectralMatching}
\title{Spectral matching}
\usage{
spectralMatching(q_dbPth, l_dbPth = NA, q_purity = NA,
  q_ppmProd = 10, q_ppmPrec = 5, q_raThres = 0, q_pol = "positive",
  q_instrumentTypes = NA, q_instrument = NA, q_sources = NA,
  q_spectraType = "av_all", q_sids = NA, q_rtrange = c(NA, NA),
  q_spectraFilter = TRUE, q_xcmsGroups = NA, l_purity = NA,
  l_ppmProd = 10, l_ppmPrec = 5, l_raThres = 0, l_pol = "positive",
  l_instrumentTypes = NA, l_instrument = NA, l_sources = NA,
  l_spectraType = NA, l_sids = NA, l_rtrange = c(NA, NA),
  l_spectraFilter = NA, l_xcmsGroups = NA, usePrecursors = TRUE,
  score_thres = 0.6, raW = 0.5, mzW = 2, rttol = NA,
  match_alg = "dpc", cores = 1, out_dir = ".", copy = FALSE)
}
\value{
list of database details and dataframe summarising the results for the xcms features
}
\description{
Perform spectral matching to spectral libraries using dot product cosine on a LC-MS/MS dataset and link to XCMS features.
}
\examples{
msmsPths <- list.files(system.file("extdata", "lcms", "mzML", package="msPurityData"), full.names = TRUE, pattern = "MSMS")
xset <- xcms::xcmsSet(msmsPths)
xset <- xcms::group(xset)
xset <- xcms::retcor(xset)
xset <- xcms::group(xset)

pa  <- purityA(msmsPths)
pa <- frag4feature(pa, xset)
pa <- filterFragSpectra(xset, allfrag=TRUE)
pa <- averageAllFragSpectra(pa)
q_dbPth <- createDatabase(pa, xset)
result <- spectralMatching(q_dbPth)
}
